* fix lint


* Fix issue with back button inside the register dialog


* Fix issue with Next button inside root-config component


* Show validation error for already existed username on UI


* Display validation error when entered passwords do not match


* Add whitespace validation for create account and login forms


* Add show/hide functionality for Create account/Login forms


* Remove error from server from UI, fix error message for the existing username


* Fix confirm password validation, restore data flow


* Remove console logs


* Fix control name


* Add filtering for standard registries and policies lists


* Add text for the case when there is no standard registries


* Fix issues, refactor code


* Fix issue when the user is unable to distinguish the schema policy based on version


* Disable action buttons when the form is invalid


* Mark required fields in forms


* Add disableClose attribute to modal dialogs, cross icon to some of them


* Update the UX of Registrant Application form


* update


* swagger


* fix PORT in use issue
fix build ussues
remove unnesessary credentials


* fix port generation
remove unnesessary credentials


* fix mongodb issue
remove unnesessary credentials


* update

* add swagger package
remove unnesessary credentials


* add swagger package
remove unnesessary credentials


* GITBOOK-362: Updated Installation guide

* update


* update


* added policy syntax highlighting


* added policy suggestion mechanism


* fix


* added policy syntax groups constant


* improving perfomance of suggestion algorithm


* added legend for suggestion configuration


* missing mint block tip


* added swagger docs


* changed suggestion algorithm and response


* added syntax highlighting to json block config


* changed suggestion to suggestions


* fix response codes


* change type to blockType


* fix comments


* added missing message memo


* update

* fix

* fix

* fix


* added auto-generated swagger docs


* fix


* fix


* update


* remove api-docs
remove unnesessary credentials


* fix


* Fix issues found during QA testing


* fix module editing in json


* fix


* fix perfomance on select block issue


* fix text


* fix text


* Create readMe.md

* Update readMe.md

* Update readMe.md

* Add files via upload

* Rename Methodology Library/CDM/readme.md to Methodology Library/CDM/CDM AR-ACM0003/readme.md

* Add files via upload

* Delete CDM.policy

* Add files via upload

* policy.ts
remove unnesessary credentials


* fix 2280
remove unnesessary credentials


* fix 2242
remove unnesessary credentials


* feat: preview and adjustments

* Fix issue with action button disabling in token dialog


* Change filtering by username to geography


* Change the style of Generate button for SR, Registrant users


* GITBOOK-364: Separated sections from Installation guide

* Add hover state for the generate button


* enable config view of published modules
remove unnesessary credentials


* fix module "save as"
remove unnesessary credentials


* merge develop
remove unnesessary credentials


* fix bugs
remove unnesessary credentials


* fix get accounts
remove unnesessary credentials


* fix status codes
remove unnesessary credentials


* added ghgp policy [skip ci]


* Updates main ICP policy to v.2.6.1.


* fix accounts
remove unnesessary credentials


* GITBOOK-366: Separating Installation methods

* fix status codes
remove unnesessary credentials


* fix permissions
remove unnesessary credentials


* GITBOOK-369: Corrected manuaaly process

* GITBOOK-371: Added Auto Suggestion docs

* GITBOOK-372: Added syntax section in themes

* docs generation test
remove unnesessary credentials


* docs generation test
remove unnesessary credentials


* fix filter in report blocks

* added NEW status for vc-document


* fix filter in report blocks


* fix invalid status codes
remove unnesessary credentials


* Solve #1695
AWS Env secret
HCP Env secret
AZURE Env secret
Documentation
solve conflicts
api-docs


* six send VP


* fix lint
remove unnesessary credentials


* update cdm policy [skip ci]


* added cdm ams policy [skip ci]


* GITBOOK-373: added messagesReportBlock

* fix status codes
remove unnesessary credentials


* fix modules
remove unnesessary credentials


* fix profiles
remove unnesessary credentials


* fix module states


* fix validation


* up hedera sdk version
add network name in logs


* fix style


* fix swagger.yaml generation


* fix swagger.yaml generation


* fix swagger.yaml generation


* Add swagger.yaml

* fix swagger.yaml generation


* fix swagger.yaml generation


* test fix swagger.yaml generation


* test fix swagger.yaml generation


* test fix swagger.yaml generation


* test fix swagger.yaml generation


* test fix swagger.yaml generation


* Add swagger.yaml

* test fix swagger.yaml generation


* [skip ci] Add swagger.yaml

* save serials

* ignore main brunch


* GITBOOK-374: Ux in selecting standards update

* fix error codes


* [skip ci] Add swagger.yaml

* add missing keywords for policy syntax


* refactor syntax groups theme storage


* update

* fix preset readonly nested schemas


* update


* auto fill suggestions config


* fix error codes


* fix


* fix permissions


* update


* fix errors


* fix errors


* fix error


* fix errors


* fix error


* fix error


* fix error


* compare modules


* [skip ci] Add swagger.yaml

* disable vault in api-tests


* update policies

* compare module


* update compare ui


* fix search btn


* fixes


* fix lint


* [skip ci] Add swagger.yaml

* fix permission


* fix permission


* fix


* #2064 fix mobile menu


* merge main


* fix


* remove latest tags


* update serials column [skip ci]


* updated Irec timestamps

* fix docker start


* updated iREC10 recipient timestamp

* GITBOOK-376: Updated iREC timestamps

* added copy button to new versions


* fix createDate\updateDate in db

* fix navigate

* fix disable btn in request block


* fix schema cache


* hotfix/form-scroll-buttons


* Update the initial isShown array value


* add swagger descriptions


* [skip ci] Add swagger.yaml

* fix suggestions on deleting policy/module


* use static swagger descriptions if need


* [skip ci] Add swagger.yaml

* bump version


* bump version


* [skip ci] Add swagger.yaml

* fix lock file


* GITBOOK-377: api-architecture-customization

---------

+++ api-gateway/src/api/service/tokens.ts
@@ -7,6 +7,7 @@ import { ServiceError } from '@helpers/service-requests-base';
 import { prepareValidationResponse } from '@middlewares/validation';
 import { Controller, Delete, Get, HttpCode, HttpException, HttpStatus, Post, Put, Req, Response } from '@nestjs/common';
 import { checkPermission } from '@auth/authorization-helper';
+import { ApiTags } from '@nestjs/swagger';
 
 /**
  * Token route
@@ -27,10 +28,12 @@ function setTokensPolicies<T>(tokens: any[], map: any[], policyId?: any, notEmpt
     for (const token of tokens) {
         token.policies = token.policies || [];
         token.policyIds = token.policyIds || [];
+        token.canDelete = true;
         for (const policyObject of map) {
             if (policyObject.tokenIds.includes(token.tokenId)) {
                 token.policies.push(`${policyObject.name} (${policyObject.version || 'DRAFT'})`);
                 token.policyIds.push(policyObject.id.toString());
+                token.canDelete = token.canDelete && policyObject.status === 'DRAFT';
             }
         }
     }
@@ -69,6 +72,7 @@ async function setDynamicTokenPolicy(tokens: any[], engineService?: PolicyEngine
 }
 
 @Controller('tokens')
+@ApiTags('tokens')
 export class TokensApi {
     @Get('/')
     @HttpCode(HttpStatus.OK)
@@ -125,7 +129,7 @@ export class TokensApi {
             return res.status(201).json(tokens);
         } catch (error) {
             new Logger().error(error, ['API_GATEWAY']);
-            throw error;
+            throw new HttpException(error.message, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
@@ -222,14 +226,21 @@ export class TokensApi {
                 throw new HttpException('Invalid creator.', HttpStatus.FORBIDDEN);
             }
 
+            const engineService = new PolicyEngine();
+            const map = await engineService.getTokensMap(user.did);
+            setTokensPolicies([tokenObject], map, undefined, false);
+
+            if (!tokenObject.canDelete) {
+                throw new HttpException('Token cannot be deleted', HttpStatus.FORBIDDEN);
+            }
             RunFunctionAsync<ServiceError>(async () => {
                 await guardians.deleteTokenAsync(tokenId, taskId);
             }, async (error) => {
                 new Logger().error(error, ['API_GATEWAY']);
-                taskManager.addError(taskId, { code: error.code || 500, message: error.message });
+                taskManager.addError(taskId, {code: error.code || 500, message: error.message});
             });
 
-            return res.status(202).send({ taskId, expectation });
+            return res.status(202).send({taskId, expectation});
         } catch (error) {
             new Logger().error(error, ['API_GATEWAY']);
             throw error;
@@ -257,7 +268,7 @@ export class TokensApi {
             if (error?.message?.toLowerCase().includes('token not found')) {
                 throw new HttpException('Token does not exist.', HttpStatus.NOT_FOUND)
             }
-            throw error;
+            throw new HttpException(error.message, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
@@ -306,7 +317,7 @@ export class TokensApi {
             if (error?.message?.toLowerCase().includes('token not found')) {
                 throw new HttpException('Token does not exist.', HttpStatus.NOT_FOUND)
             }
-            throw error;
+            throw new HttpException(error.message, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
@@ -584,7 +595,7 @@ export class TokensApi {
             if (error?.message?.toLowerCase().includes('token not found')) {
                 throw new HttpException('Token not registered', HttpStatus.NOT_FOUND);
             }
-            throw error
+            throw new HttpException(error.message, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
@@ -610,7 +621,7 @@ export class TokensApi {
             if (error?.message?.toLowerCase().includes('token not found')) {
                 throw new HttpException('Token not registered', HttpStatus.NOT_FOUND);
             }
-            throw error;
+            throw new HttpException(error.message, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
@@ -642,6 +653,7 @@ export class TokensApi {
     @Put('/push/:tokenId/:username/unfreeze')
     @HttpCode(HttpStatus.ACCEPTED)
     async unfreezeTokenAsync(@Req() req, @Response() res): Promise<any> {
+        await checkPermission(UserRole.STANDARD_REGISTRY)(req.user);
         const taskManager = new TaskManager();
         const { taskId, expectation } = taskManager.start('Unfreeze Token');
